acsess handeling v1

+++ main.py
@@ -123,9 +123,13 @@ async def delete_user(user_id: int, db: Session = Depends(get_db), token: str = 
     # بررسی نقش کاربر
     if user.role.value != "admin":
         raise HTTPException(status_code=403, detail="شما قادر به انجام این عملیات نیستید.")
+
     user_model = db.query(models.User).filter(models.User.id == user_id).first()
     if user_model is None:
         raise HTTPException(status_code=404, detail="کاربر یافت نشد .")
+    users_parent_id_is_not_none = db.query(models.User).filter(models.User.ParentId == user_id).first()
+    if users_parent_id_is_not_none:
+        raise HTTPException(status_code=400, detail='سرپرست داری زیر مجموعه میباشد .')
     db.query(models.User).filter(models.User.id == user_id).delete()
     db.commit()
     raise HTTPException(status_code=200, detail='حذف با موفقیت انجام شد .')
@@ -161,7 +165,9 @@ async def update_user(user_id: int, user_update: UserUpdate, db: Session = Depen
 
     if user_model is None:
         raise HTTPException(status_code=404, detail="کابر یافت نشد.")
-
+    users_parent_id_is_not_none = db.query(models.User).filter(models.User.ParentId == user.id)
+    if users_parent_id_is_not_none:
+        raise HTTPException(status_code=400, detail='سرپرست داری زیر مجموعه میباشد .')
     if user_update.UserName:
         user_model.UserName = user_update.UserName
     if user_update.Name:
